overseer/requirements.txt
tornado
docker
sockjs-tornado
pyzmq
boto3
egenix-mx-base
requests
ruamel.yaml
